Improve the installation of Bulma and the theme
This simplifies how Bulma is wired up and imported into components.  Bulma doesn't natively support TypeScript yet (will soon with the pending 4.0 release!) so there is more room for improvement here in the near future.  This should makes things a bit cleaner and maintainable.  For instance, I was able to remove the compulsory `_variables.sass` file that it was dependent on previously.