Add wave background behind "about me" section
Using VantaJs, I added a wave background behind the "about me" section.  It's themed to match.  Unfortunately, Vanta doesn't have TS support yet.  The implementation was largely copied and pasted with a few changes:

1. Define more settings for the wave effect
2. Add `three` as a dependency instead of using a CDN
3. Fixed what TS errors were in my control

Following the SASS 7-1 architecture, a new "components" folder was created for the CSS to make the background transparent for the `Section` component.