Fix bug on mobile that made the layout wonky
Viewing the site on mobile caused an issue where sections were getting chopped over and/or overlapping.  Using a `Hero` component to define the section has built-in support for making a section full-width and full-height while keeping the content centered.  This was what I was originally trying to accomplish in my previous attempt.  This fix is nice because it's more declarative in what I'm trying to do and simpler.