Add navbar to page
The Bulma navbar has been wired up to work with `react-scroll`.  A custom hook (`useOffset`) was created to compute the height of the navbar at runtime.  This was necessary because `react-scroll` doesn't know to take in the height of the navbar when scrolling to different components causing it to "over scroll" by the height of the navbar.