Fix for hamburger menu not working on mobile
Bulma doesn't ship with _any_ JS, so clicking on the hamburger menu did nothing.  This adds support to add the `is-active` class to the appropriate elements when the hamburger menu is clicked on.  Likewise, it'll remove the `is-active` class when the "X" is clicked on or when an item in the menu is clicked on so that the menu will collapse.